update the utils and extract reusable code
