+++ src/tabs/upload/lib/utils.ts
@@ -1,45 +1,9 @@
 /**
- * Validates an array of URL strings and separates them into valid and invalid arrays
+ * Upload-tab constants: API endpoints and batch dropdown value.
  */
-export function validateUrls(urlLines: string[]): {
-  valid: string[];
-  invalid: string[];
-} {
-  const valid: string[] = [];
-  const invalid: string[] = [];
 
-  for (const url of urlLines) {
-    try {
-      new URL(url);
-      valid.push(url);
-    } catch {
-      invalid.push(url);
-    }
-  }
-
-  return { valid, invalid };
-}
-
-/**
- * Extracts company name from a URL by taking the first part of the hostname
- * Falls back to "Unknown" if parsing fails
- */
-export function extractCompanyFromUrl(url: string): string {
-  try {
-    return new URL(url).hostname.split(".")[0] || "Unknown";
-  } catch {
-    return "Unknown";
-  }
-}
-
-/**
- * API endpoint for PDF parsing
- */
 export const PARSE_PDF_API_ENDPOINT = "/api/queues/parsePdf";
-
-/**
- * API endpoint to list existing batch IDs (for dropdown)
- */
 export const BATCHES_API_ENDPOINT = "/api/processes/batches";
 
-export { DEFAULT_RUN_ONLY } from "@/lib/run-only-workers";
+/** Value for "new batch" in the batch dropdown; when selected, customBatchName is used. */
+export const NEW_BATCH_DROPDOWN_VALUE = "__new__";
